requirements.txt
tensorflow
fastapi
uvicorn
streamlit
pillow
numpy
pytz            # time zone management
# tests
httpx           # HTTP client
pytest-asyncio  # asynchronous I/O support for pytest
